Added and modified a few files, inclusive of MainApp, Reservation, Table and ReservationMenuUI; added a new CSV reservations.csv for saving purposes and testing purposes. A new helper class is also added to aid in formatting of dates from String to Calendar, Calendar to String, etc.
Added stub functions for Reservation functions, added default csv file for storing Reservations. Added new CSV helper for Reservations.
DateTimeFormatHelper class will help to convert date and times from String to Calendar and vice versa, as well as comparing and retrieving today's date, for ease of Reservations, and Invoice as Invoice needs to keep track of bill date.

REMOVED Seat class as it is unnecessary - no useful attributes have been defined for Seats as we do not need to keep track of who is at which Seat, and etc.